// src/electron/types/helper.ts
export enum WsEventType {
  SYSTEM_MESSAGE = "system:message",
  SYSTEM_MESSAGE_START = "system:message:start",
  SYSTEM_MESSAGE_CHUNK = "system:message:chunk",
  SYSTEM_MESSAGE_END = "system:message:end",
  SYSTEM_AUDIO_CHUNK = "system:audio:chunk",
  SYSTEM_QUOTA_ALERT = "system:exhausted:alert",
  SYSTEM_QUOTA_EXHAUSTED = "system:exhausted:exhaust",
  SYSTEM_PING = "system:ping",
  SYSTEM_ERROR = "system:error",
  AUTH_EXPIRED = "auth:expired",
  USER_MESSAGE = "user:message",
  USER_IMAGE = "user:image",
  USER_TRANSCRIPTION = "user:transcription",
  USER_AUDIO_CHUNK = "user:audio:chunk",
  USER_MEDIA = "user:media",
  USER_SESSION_UPDATE = "user:session:update",
}